Merge alibaba/canal master changes

+++ canal-admin/canal-admin-server/src/main/java/com/alibaba/otter/canal/admin/service/impl/NodeServerServiceImpl.java
@@ -1,24 +1,29 @@
 package com.alibaba.otter.canal.admin.service.impl;
 
+import io.ebean.Query;
+
+import java.security.NoSuchAlgorithmException;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.concurrent.ExecutionException;
-import java.util.concurrent.ExecutorService;
-import java.util.concurrent.Executors;
 import java.util.concurrent.Future;
+import java.util.concurrent.TimeUnit;
+import java.util.concurrent.TimeoutException;
 
 import org.apache.commons.lang.StringUtils;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
 
+import com.alibaba.otter.canal.admin.common.TemplateConfigLoader;
+import com.alibaba.otter.canal.admin.common.Threads;
 import com.alibaba.otter.canal.admin.common.exception.ServiceException;
-import com.alibaba.otter.canal.admin.jmx.CanalServerMXBean;
-import com.alibaba.otter.canal.admin.jmx.JMXConnection;
+import com.alibaba.otter.canal.admin.connector.AdminConnector;
+import com.alibaba.otter.canal.admin.connector.SimpleAdminConnectors;
+import com.alibaba.otter.canal.admin.model.CanalConfig;
+import com.alibaba.otter.canal.admin.model.CanalInstanceConfig;
 import com.alibaba.otter.canal.admin.model.NodeServer;
+import com.alibaba.otter.canal.admin.model.Pager;
 import com.alibaba.otter.canal.admin.service.NodeServerService;
-
-import io.ebean.Query;
+import com.alibaba.otter.canal.protocol.SecurityUtil;
 
 /**
  * 节点信息业务层
@@ -29,19 +34,30 @@ import io.ebean.Query;
 @Service
 public class NodeServerServiceImpl implements NodeServerService {
 
-    private static final Logger logger = LoggerFactory.getLogger(NodeServerServiceImpl.class);
-
     public void save(NodeServer nodeServer) {
         int cnt = NodeServer.find.query()
             .where()
             .eq("ip", nodeServer.getIp())
-            .eq("port", nodeServer.getPort())
+            .eq("adminPort", nodeServer.getAdminPort())
             .findCount();
         if (cnt > 0) {
             throw new ServiceException("节点信息已存在");
         }
 
         nodeServer.save();
+
+        if (nodeServer.getClusterId() == null) { // 单机模式
+            CanalConfig canalConfig = new CanalConfig();
+            canalConfig.setServerId(nodeServer.getId());
+            String configTmp = TemplateConfigLoader.loadCanalConfig();
+            canalConfig.setContent(configTmp);
+            try {
+                String contentMd5 = SecurityUtil.md5String(canalConfig.getContent());
+                canalConfig.setContentMd5(contentMd5);
+            } catch (NoSuchAlgorithmException e) {
+            }
+            canalConfig.save();
+        }
     }
 
     public NodeServer detail(Long id) {
@@ -52,25 +68,39 @@ public class NodeServerServiceImpl implements NodeServerService {
         int cnt = NodeServer.find.query()
             .where()
             .eq("ip", nodeServer.getIp())
-            .eq("port", nodeServer.getPort())
+            .eq("adminPort", nodeServer.getAdminPort())
             .ne("id", nodeServer.getId())
             .findCount();
         if (cnt > 0) {
             throw new ServiceException("节点信息已存在");
         }
 
-        nodeServer.update("name", "ip", "port", "port2");
+        nodeServer.update("name", "ip", "adminPort", "tcpPort", "metricPort", "clusterId");
     }
 
     public void delete(Long id) {
         NodeServer nodeServer = NodeServer.find.byId(id);
         if (nodeServer != null) {
+            // 判断是否存在实例
+            int cnt = CanalInstanceConfig.find.query().where().eq("serverId", nodeServer.getId()).findCount();
+            if (cnt > 0) {
+                throw new ServiceException("当前Server下存在Instance配置, 无法删除");
+            }
+
+            // 同时删除配置
+            CanalConfig canalConfig = CanalConfig.find.query().where().eq("serverId", id).findOne();
+            if (canalConfig != null) {
+                canalConfig.delete();
+            }
+
             nodeServer.delete();
         }
     }
 
-    public List<NodeServer> findList(NodeServer nodeServer) {
+    private Query<NodeServer> getBaseQuery(NodeServer nodeServer) {
         Query<NodeServer> query = NodeServer.find.query();
+        query.fetch("canalCluster", "name").setDisableLazyLoading(true);
+
         if (nodeServer != null) {
             if (StringUtils.isNotEmpty(nodeServer.getName())) {
                 query.where().like("name", "%" + nodeServer.getName() + "%");
@@ -78,51 +108,66 @@ public class NodeServerServiceImpl implements NodeServerService {
             if (StringUtils.isNotEmpty(nodeServer.getIp())) {
                 query.where().eq("ip", nodeServer.getIp());
             }
-        }
-        query.order().asc("id");
-        List<NodeServer> nodeServers = query.findList();
-        if (nodeServers.isEmpty()) {
-            return nodeServers;
+            if (nodeServer.getClusterId() != null) {
+                if (nodeServer.getClusterId() == -1) {
+                    query.where().isNull("clusterId");
+                } else {
+                    query.where().eq("clusterId", nodeServer.getClusterId());
+                }
+            }
         }
 
-        ExecutorService executorService = Executors.newFixedThreadPool(nodeServers.size());
+        return query;
+    }
+
+    public List<NodeServer> findAll(NodeServer nodeServer) {
+        Query<NodeServer> query = getBaseQuery(nodeServer);
+        query.order().asc("id");
+        return query.findList();
+    }
+
+    public Pager<NodeServer> findList(NodeServer nodeServer, Pager<NodeServer> pager) {
+
+        Query<NodeServer> query = getBaseQuery(nodeServer);
+        Query<NodeServer> queryCnt = query.copy();
+
+        int count = queryCnt.findCount();
+        pager.setCount((long) count);
+
+        List<NodeServer> nodeServers = query.order()
+            .asc("id")
+            .setFirstRow(pager.getOffset().intValue())
+            .setMaxRows(pager.getSize())
+            .findList();
+        pager.setItems(nodeServers);
+
+        if (nodeServers.isEmpty()) {
+            return pager;
+        }
+
         List<Future<Boolean>> futures = new ArrayList<>(nodeServers.size());
         // get all nodes status
         for (NodeServer ns : nodeServers) {
-            futures.add(executorService.submit(() -> {
-                int status = -1;
-                JMXConnection jmxConnection = new JMXConnection(ns.getIp(), ns.getPort());
-                try {
-                    CanalServerMXBean canalServerMXBean = jmxConnection.getCanalServerMXBean();
-                    status = canalServerMXBean.getStatus();
-                } catch (Exception e) {
-                    logger.warn(e.getMessage());
-                } finally {
-                    jmxConnection.close();
-                }
-                ns.setStatus(status);
-                return status != -1;
+            futures.add(Threads.executorService.submit(() -> {
+                boolean status = SimpleAdminConnectors.execute(ns.getIp(), ns.getAdminPort(), AdminConnector::check);
+                ns.setStatus(status ? "1" : "0");
+                return !status;
             }));
         }
         futures.forEach(f -> {
             try {
-                f.get();
-            } catch (InterruptedException | ExecutionException e) {
+                f.get(3, TimeUnit.SECONDS);
+            } catch (TimeoutException | InterruptedException | ExecutionException e) {
                 // ignore
             }
         });
 
-        executorService.shutdownNow();
-
-        return nodeServers;
+        return pager;
     }
 
     public int remoteNodeStatus(String ip, Integer port) {
-        Integer resutl = JMXConnection.execute(ip, port, CanalServerMXBean::getStatus);
-        if (resutl == null) {
-            resutl = -1;
-        }
-        return resutl;
+        boolean result = SimpleAdminConnectors.execute(ip, port, AdminConnector::check);
+        return result ? 1 : 0;
     }
 
     public String remoteCanalLog(Long id) {
@@ -130,7 +175,9 @@ public class NodeServerServiceImpl implements NodeServerService {
         if (nodeServer == null) {
             return "";
         }
-        return JMXConnection.execute(nodeServer.getIp(), nodeServer.getPort(), CanalServerMXBean::canalLog);
+        return SimpleAdminConnectors.execute(nodeServer.getIp(),
+            nodeServer.getAdminPort(),
+            adminConnector -> adminConnector.canalLog(100));
     }
 
     public boolean remoteOperation(Long id, String option) {
@@ -138,18 +185,18 @@ public class NodeServerServiceImpl implements NodeServerService {
         if (nodeServer == null) {
             return false;
         }
-        Boolean resutl = null;
+        Boolean result = null;
         if ("start".equals(option)) {
-            resutl = JMXConnection.execute(nodeServer.getIp(), nodeServer.getPort(), CanalServerMXBean::start);
+            result = SimpleAdminConnectors.execute(nodeServer.getIp(), nodeServer.getAdminPort(), AdminConnector::start);
         } else if ("stop".equals(option)) {
-            resutl = JMXConnection.execute(nodeServer.getIp(), nodeServer.getPort(), CanalServerMXBean::stop);
+            result = SimpleAdminConnectors.execute(nodeServer.getIp(), nodeServer.getAdminPort(), AdminConnector::stop);
         } else {
             return false;
         }
 
-        if (resutl == null) {
-            resutl = false;
+        if (result == null) {
+            result = false;
         }
-        return resutl;
+        return result;
     }
 }
